Adding Product Entity

+++ P03-ReviewsAPI/src/main/resources/db/migration/V1_0__Create_table_product.sql
@@ -1,5 +1,5 @@
 CREATE TABLE product (
-     product_id MEDIUMINT NOT NULL AUTO_INCREMENT,
+     product_id INT NOT NULL AUTO_INCREMENT,
      title CHAR(255) NOT NULL,
      description VARCHAR(1000) NOT NULL,
      created_time TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
