+++ P03-ReviewsAPI/src/main/resources/db/migration/V1_1__Create_table_review.sql
@@ -1,6 +1,6 @@
 CREATE TABLE review (
-     review_id MEDIUMINT NOT NULL AUTO_INCREMENT,
-     product_id MEDIUMINT NOT NULL,
+     review_id INT NOT NULL AUTO_INCREMENT,
+     product_id INT NOT NULL,
      title CHAR(255) NOT NULL,
      cutomer CHAR(255) NOT NULL,
      description VARCHAR(1000) NOT NULL,
